fix: In Linux CL/GL sharing, handle failed symbols lookup
In some cases we endup with GL being found but neither GLX or EGL
found and this currently leads to crash.

So to better handle this case (and others), we keep track if we
found symbols or not.

If we're asked if CL/GL sharing is supported we check the load worked.

If it's during an actual context creation we check what we were given
match what we could load and fail context property finalization if not.

If it's during an "info" request where we don't have any properties
yet, we return yes if either GLX or EGL was found.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>