fix: Don't disable bindless allocation using an undocumented env var
Bindless allocation is a core functionality. Having obscure mechanisms
disabling it at runtime leads to unwanted side effects, leading to
application issues, for example on applications trying to use SYCL
bindless textures, that are supposed to be available unconditionally on
DG2 and newer (when texturing is supported).
This is what is assumed here:
https://github.com/oneapi-src/unified-runtime/blob/dd7d5c6256f7ec90be9753e6e5a24d06a6b36fd6/source/adapters/level_zero/device.cpp#L1095

This reverts commit 81644a46cc466a892270d1688f4da24c2165900c.

Signed-off-by: Xavier Hallade <me@ph0b.com>