feature: unify local memory size getter for i915 and xe
Previously i915 was using a sysfs node to query the amount of local
memory of a tile. For xe such sysfs node does not exist so the function
was useless.

Instead of a sysfs node, leverage the local memory size present in
MemoryInfo (queried at init-stage) so that the output is now the same
for both i915 and xe.

Related-To: NEO-11127
Signed-off-by: Maciej Bielski <maciej.bielski@intel.com>