fix: avoid race when evicting resources
Related-To: NEO-13843

DrmMemoryOperationsHandler::mutex is used for residency
handling.
However, this mutex is not being locked during eviction when vm bind
failed.
This is causing races for example when ULLS controller wants to make
resources resident, and user thread wants to evict
them.
Fix is to explicitly obtain lock in problematic path.

Signed-off-by: Szymon Morek <szymon.morek@intel.com>