refactor: remove misleading std::move
The `surfStateInfo` argument has a `const SurfStateInHeapInfo &` type.
Thus the result of `std::move()` is
`std::remove_reference<const SurfStateHeapInfo &>::type &&`.
This effectively boils down to `const SurfStateHeapInfo &` (the L-value
reference has precedence), which effectively is passed to
`.push_back(const SurfStateHeapInfo &)`.
This is the same outcome as if the `std::move()` is not present at all.

Related-To: NEO-15630

Signed-off-by: Maciej Bielski <maciej.bielski@intel.com>